+++ packages/eas-cli/src/credentials/ios/api/graphql/queries/AppleDistributionCertificateQuery.ts
@@ -5,13 +5,14 @@ import gql from 'graphql-tag';
 import { ExpoGraphqlClient } from '../../../../../commandUtils/context/contextUtils/createGraphqlClient';
 import { withErrorHandlingAsync } from '../../../../../graphql/client';
 import {
-  AppleDistributionCertificateByAccountQuery,
   AppleDistributionCertificateByAppQuery,
   AppleDistributionCertificateFragment,
+  AppleDistributionCertificatesPaginatedByAccountQuery,
   IosDistributionType,
 } from '../../../../../graphql/generated';
 import { AppleDistributionCertificateFragmentNode } from '../../../../../graphql/types/credentials/AppleDistributionCertificate';
 import { AppleTeamFragmentNode } from '../../../../../graphql/types/credentials/AppleTeam';
+import { Connection, QueryParams, fetchEntireDatasetAsync } from '../../../../../utils/relay';
 
 export const AppleDistributionCertificateQuery = {
   async getForAppAsync(
@@ -81,17 +82,65 @@ export const AppleDistributionCertificateQuery = {
     graphqlClient: ExpoGraphqlClient,
     accountName: string
   ): Promise<AppleDistributionCertificateFragment[]> {
+    const paginatedGetterAsync = async (
+      relayArgs: QueryParams
+    ): Promise<Connection<AppleDistributionCertificateFragment>> => {
+      return await AppleDistributionCertificateQuery.getAllForAccountPaginatedAsync(
+        graphqlClient,
+        accountName,
+        relayArgs
+      );
+    };
+    return await fetchEntireDatasetAsync({
+      paginatedGetterAsync,
+      progressBarLabel: 'fetching Apple Distribution Certificates...',
+    });
+  },
+  async getAllForAccountPaginatedAsync(
+    graphqlClient: ExpoGraphqlClient,
+    accountName: string,
+    {
+      after,
+      first,
+      before,
+      last,
+    }: { after?: string; first?: number; before?: string; last?: number }
+  ): Promise<
+    AppleDistributionCertificatesPaginatedByAccountQuery['account']['byName']['appleDistributionCertificatesPaginated']
+  > {
     const data = await withErrorHandlingAsync(
       graphqlClient
-        .query<AppleDistributionCertificateByAccountQuery>(
+        .query<AppleDistributionCertificatesPaginatedByAccountQuery>(
           gql`
-            query AppleDistributionCertificateByAccountQuery($accountName: String!) {
+            query AppleDistributionCertificatesPaginatedByAccountQuery(
+              $accountName: String!
+              $after: String
+              $first: Int
+              $before: String
+              $last: Int
+            ) {
               account {
                 byName(accountName: $accountName) {
                   id
-                  appleDistributionCertificates {
-                    id
-                    ...AppleDistributionCertificateFragment
+                  appleDistributionCertificatesPaginated(
+                    after: $after
+                    first: $first
+                    before: $before
+                    last: $last
+                  ) {
+                    edges {
+                      cursor
+                      node {
+                        id
+                        ...AppleDistributionCertificateFragment
+                      }
+                    }
+                    pageInfo {
+                      hasNextPage
+                      hasPreviousPage
+                      startCursor
+                      endCursor
+                    }
                   }
                 }
               }
@@ -100,6 +149,10 @@ export const AppleDistributionCertificateQuery = {
           `,
           {
             accountName,
+            after,
+            first,
+            before,
+            last,
           },
           {
             additionalTypenames: ['AppleDistributionCertificate'],
@@ -107,6 +160,6 @@ export const AppleDistributionCertificateQuery = {
         )
         .toPromise()
     );
-    return data.account.byName.appleDistributionCertificates;
+    return data.account.byName.appleDistributionCertificatesPaginated;
   },
 };